pyproject.toml
[tool.poetry]
name = "fs-starter"
version = "0.1.0"
description = ""
authors = ["Dave Kuhn <dave@stack.biz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = {extras = ["redis"], version = "^4.2.5"}
celery = {extras = ["redis"], version = "^5.3.4"}
django-celery-results = "^2.5.1"
psycopg2 = "^2.9.8"
django-environ = "^0.11.2"
redis = {version = "^4.6.0", extras = ["hiredis"]}
djangorestframework = "^3.14.0"
django-timescaledb = "^0.2.13"
django-cors-headers = "^4.2.0"
django-celery-beat = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

